feat(KSTIP-1246): link of gesuchdokumente is now in gesuchtranche

+++ backend/src/main/java/ch/dvbern/stip/api/dokument/service/GesuchDokumentService.java
@@ -16,7 +16,9 @@ import ch.dvbern.stip.api.dokument.type.DokumentTyp;
 import ch.dvbern.stip.api.dokument.type.Dokumentstatus;
 import ch.dvbern.stip.api.dokument.type.DokumentstatusChangeEvent;
 import ch.dvbern.stip.api.gesuch.entity.Gesuch;
+import ch.dvbern.stip.api.gesuch.entity.GesuchTranche;
 import ch.dvbern.stip.api.gesuch.repo.GesuchRepository;
+import ch.dvbern.stip.api.gesuch.repo.GesuchTrancheRepository;
 import ch.dvbern.stip.api.gesuch.type.Gesuchstatus;
 import ch.dvbern.stip.generated.dto.DokumentDto;
 import ch.dvbern.stip.generated.dto.GesuchDokumentAblehnenRequestDto;
@@ -49,24 +51,25 @@ public class GesuchDokumentService {
     private final DokumentRepository dokumentRepository;
     private final GesuchDokumentRepository gesuchDokumentRepository;
     private final GesuchRepository gesuchRepository;
+    private final GesuchTrancheRepository gesuchTrancheRepository;
     private final S3AsyncClient s3;
     private final ConfigService configService;
     private final DokumentstatusService dokumentstatusService;
 
     @Transactional
     public DokumentDto uploadDokument(
-        final UUID gesuchId,
+        final UUID gesuchTrancheId,
         final DokumentTyp dokumentTyp,
         final FileUpload fileUpload,
         final String objectId
     ) {
-        Gesuch gesuch = gesuchRepository.findByIdOptional(gesuchId).orElseThrow(NotFoundException::new);
+        GesuchTranche gesuchTranche = gesuchTrancheRepository.findByIdOptional(gesuchTrancheId).orElseThrow(NotFoundException::new);
         GesuchDokument gesuchDokument =
-            gesuchDokumentRepository.findByGesuchAndDokumentType(gesuch.getId(), dokumentTyp).orElseGet(
-                () -> createGesuchDokument(gesuch, dokumentTyp)
+            gesuchDokumentRepository.findByGesuchTrancheAndDokumentType(gesuchTranche.getId(), dokumentTyp).orElseGet(
+                () -> createGesuchDokument(gesuchTranche, dokumentTyp)
             );
         Dokument dokument = new Dokument();
-        dokument.setGesuchDokument(gesuchDokument);
+        dokument.getGesuchDokumente().add(gesuchDokument);
         dokument.setFilename(fileUpload.fileName());
         dokument.setObjectId(objectId);
         dokument.setFilesize(String.valueOf(fileUpload.size()));
@@ -77,9 +80,9 @@ public class GesuchDokumentService {
     }
 
     @Transactional
-    public List<DokumentDto> findGesuchDokumenteForTyp(final UUID gesuchId, final DokumentTyp dokumentTyp) {
+    public List<DokumentDto> findGesuchDokumenteForTyp(final UUID gesuchTrancheId, final DokumentTyp dokumentTyp) {
         GesuchDokument gesuchDokument =
-            gesuchDokumentRepository.findByGesuchAndDokumentType(gesuchId, dokumentTyp).orElse(null);
+            gesuchDokumentRepository.findByGesuchTrancheAndDokumentType(gesuchTrancheId, dokumentTyp).orElse(null);
         if (gesuchDokument == null) {
             return new ArrayList<>();
         }
@@ -93,20 +96,17 @@ public class GesuchDokumentService {
         return Optional.ofNullable(dokumentMapper.toDto(dokument));
     }
 
-    @Transactional(TxType.REQUIRES_NEW)
-    public void deleteAllDokumentForGesuchInRepository(final UUID gesuchId) {
-        gesuchDokumentRepository.findAllForGesuch(gesuchId)
-            .forEach(gesuchDokumentRepository::delete);
+    public void removeAllGesuchDokumentsForGesuch(final UUID gesuchId) {
+        gesuchRepository.requireById(gesuchId).getGesuchTranchen().forEach(
+            gesuchTranche -> removeAllDokumentsForGesuchTranche(gesuchTranche.getId())
+        );
     }
 
-    @Transactional
-    public List<String> getAllDokumentObjectIdsForGesuch(final UUID gesuchId) {
-        return gesuchDokumentRepository.findAllForGesuch(gesuchId)
-            .map(GesuchDokument::getDokumente)
-            .flatMap(dokuments -> dokuments
-                .stream()
-                .map(Dokument::getObjectId)
-            ).toList();
+    @Transactional(TxType.REQUIRES_NEW)
+    public void removeAllDokumentsForGesuchTranche(final UUID gesuchTrancheId) {
+        gesuchDokumentRepository.findAllForGesuchTranche(gesuchTrancheId).forEach(
+            gesuchDokument -> removeGesuchDokument(gesuchDokument.getId())
+        );
     }
 
     private static void gesuchstatusIsNotOrElseThrow(final Gesuch gesuch, final Gesuchstatus statusToVerify) {
@@ -120,7 +120,7 @@ public class GesuchDokumentService {
     @Transactional
     public void gesuchDokumentAblehnen(final UUID gesuchDokumentId, final GesuchDokumentAblehnenRequestDto dto) {
         final var gesuchDokument = gesuchDokumentRepository.requireById(gesuchDokumentId);
-        gesuchstatusIsNotOrElseThrow(gesuchDokument.getGesuch(), Gesuchstatus.IN_BEARBEITUNG_SB);
+        gesuchstatusIsNotOrElseThrow(gesuchDokument.getGesuchTranche().getGesuch(), Gesuchstatus.IN_BEARBEITUNG_SB);
         dokumentstatusService.triggerStatusChangeWithComment(
             gesuchDokument,
             DokumentstatusChangeEvent.ABGELEHNT,
@@ -131,16 +131,11 @@ public class GesuchDokumentService {
     @Transactional
     public void gesuchDokumentAkzeptieren(final UUID gesuchDokumentId) {
         final var gesuchDokument = gesuchDokumentRepository.requireById(gesuchDokumentId);
-        gesuchstatusIsNotOrElseThrow(gesuchDokument.getGesuch(), Gesuchstatus.IN_BEARBEITUNG_SB);
+        gesuchstatusIsNotOrElseThrow(gesuchDokument.getGesuchTranche().getGesuch(), Gesuchstatus.IN_BEARBEITUNG_SB);
         dokumentstatusService.triggerStatusChange(
             gesuchDokument,
             DokumentstatusChangeEvent.AKZEPTIERT
         );
-    }
-
-    public void deleteAllDokumentForGesuch(final UUID gesuchId) {
-        executeDeleteDokumentsFromS3(getAllDokumentObjectIdsForGesuch(gesuchId));
-        deleteAllDokumentForGesuchInRepository(gesuchId);
     }
 
     public CompletableFuture<DeleteObjectsResponse> deleteDokumentsFromS3Blocking(final List<String> objectIds) {
@@ -173,9 +168,50 @@ public class GesuchDokumentService {
     public String deleteDokument(final UUID dokumentId) {
         Dokument dokument = dokumentRepository.findByIdOptional(dokumentId).orElseThrow(NotFoundException::new);
         final var dokumentObjectId = dokument.getObjectId();
-        dokumentRepository.delete(dokument);
-        gesuchDokumentRepository.dropGesuchDokumentIfNoDokumente(dokument.getGesuchDokument().getId());
+        for (final var gesuchDokument : dokument.getGesuchDokumente()) {
+            gesuchDokument.getDokumente().remove(dokument);
+            gesuchDokumentRepository.dropGesuchDokumentIfNoDokumente(gesuchDokument.getId());
+        }
+        if (dokument.getGesuchDokumente().isEmpty()) {
+            dokumentRepository.delete(dokument);
+        }
+
         return dokumentObjectId;
+    }
+
+    @Transactional(TxType.REQUIRES_NEW)
+    public void removeDokument(final UUID dokumentId) {
+        Dokument dokument = dokumentRepository.findByIdOptional(dokumentId).orElseThrow(NotFoundException::new);
+        final var dokumentObjectId = dokument.getObjectId();
+        for (final var gesuchDokument : dokument.getGesuchDokumente()) {
+            gesuchDokument.getDokumente().remove(dokument);
+            gesuchDokumentRepository.dropGesuchDokumentIfNoDokumente(gesuchDokument.getId());
+        }
+        if (dokument.getGesuchDokumente().isEmpty()) {
+            dokumentRepository.delete(dokument);
+            executeDeleteDokumentsFromS3(List.of(dokumentObjectId));
+        }
+    }
+
+    @Transactional(TxType.REQUIRES_NEW)
+    public void removeGesuchDokument(final UUID gesuchDokumentId) {
+        final var gesuchDokument = gesuchDokumentRepository.requireById(gesuchDokumentId);
+        final var dokuments = gesuchDokument.getDokumente();
+
+        List<String> dokumentObjectIds = new ArrayList<>();
+        for (final var dokument : dokuments) {
+            gesuchDokument.getDokumente().remove(dokument);
+            if (dokument.getGesuchDokumente().isEmpty()) {
+                dokumentRepository.delete(dokument);
+                dokumentObjectIds.add(dokument.getObjectId());
+            }
+        }
+        if (gesuchDokument.getDokumente().isEmpty()) {
+            gesuchDokumentRepository.delete(gesuchDokument);
+        }
+        if (!dokumentObjectIds.isEmpty()) {
+            executeDeleteDokumentsFromS3(dokumentObjectIds);
+        }
     }
 
     @Transactional
@@ -260,8 +296,8 @@ public class GesuchDokumentService {
             .build();
     }
 
-    private GesuchDokument createGesuchDokument(final Gesuch gesuch, final DokumentTyp dokumentTyp) {
-        GesuchDokument gesuchDokument = new GesuchDokument().setGesuch(gesuch).setDokumentTyp(dokumentTyp);
+    private GesuchDokument createGesuchDokument(final GesuchTranche gesuchTranche, final DokumentTyp dokumentTyp) {
+        GesuchDokument gesuchDokument = new GesuchDokument().setGesuchTranche(gesuchTranche).setDokumentTyp(dokumentTyp);
         gesuchDokumentRepository.persist(gesuchDokument);
         return gesuchDokument;
     }
